pyproject.toml
[project]
name = "groovebox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.8",
    "django-crispy-forms>=2.5",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
]
